+++ lib/nice_partials/monkey_patch.rb
@@ -29,31 +29,62 @@ end
 module NicePartials::RenderingWithAutoContext
   ActionView::Base.prepend self
 
-  def partial
-    @partial ||= nice_partial
+  def __partials
+    @__partials ||= NicePartials::Partial::Stack.new
   end
+  delegate :partial, to: :__partials
 
-  def render(options = {}, locals = {}, &block)
-    _partial = @partial
+  # Overrides `ActionView::Helpers::RenderingHelper#render` to push a new `nice_partial`
+  # on the stack, so rendering has a fresh `partial` to store content in.
+  def render(*)
+    __partials.prepend nice_partial
     super
   ensure
-    @partial = _partial
+    __partials.shift
   end
 
+  # Since Action View passes any `yield`s in partials through `_layout_for`, we
+  # override `_layout_for` to detects if it's a capturing yield and append the
+  # current partial to the arguments.
+  #
+  # So `render … do |some_object|` can become `render … do |some_object, partial|`
+  # without needing to find and update the inner `yield some_object` call.
   def _layout_for(*arguments, &block)
     if block && !arguments.first.is_a?(Symbol)
-      partial.capture(*arguments, &block)
+      capture_with_outer_partial_access(*arguments, &block)
     else
       super
     end
+  end
+
+  # Reverts `partial` to return the outer partial before the `render` call started.
+  #
+  # So we don't clobber the `partial` shown here:
+  #
+  #   <%= render "card" do |inner_partial| %>
+  #     <% inner_partial.content_for :title, partial.content_for(:title) %>
+  #   <% end %>
+  #
+  # Note: this happens because the `@partial` instance variable is shared between all
+  # `render` calls since rendering happens in one `ActionView::Base` instance.
+  def capture_with_outer_partial_access(*arguments, &block)
+    __partials.locate_previous
+    __partials.first.capture(*arguments, &block)
+  ensure
+    __partials.reset_locator
   end
 end
 
 module NicePartials::PartialRendering
   ActionView::PartialRenderer.prepend self
 
+  # Automatically captures the `block` in case the partial has no manual capturing `yield` call.
+  #
+  # This manual equivalent would be inserting this:
+  #
+  #   <% yield partial %>
   def render_partial_template(view, locals, template, layout, block)
-    view.partial.capture(&block) if block && !template.has_capturing_yield?
+    view.capture_with_outer_partial_access(&block) if block && !template.has_capturing_yield?
     super
   end
 end
@@ -72,6 +103,6 @@ module NicePartials::CapturingYieldDetection
   # Note: `<%= yield %>` becomes `yield :layout` with no `render` `block`, though this method assumes a block is passed.
   def has_capturing_yield?
     defined?(@has_capturing_yield) ? @has_capturing_yield :
-      @has_capturing_yield = source.match?(/\byield[\(? ]+(%>|[^:])/)
+      @has_capturing_yield = source.match?(/[^\.\b]yield[\(? ]+(%>|[^:])/)
   end
 end